chore(frontend): QuestionBox component
